+++ mathics/builtin/atomic/numbers.py
@@ -27,7 +27,6 @@ from functools import lru_cache
 from mathics.builtin.base import Builtin, Predefined
 from mathics.core.convert import from_sympy
 from mathics.core.evaluators import apply_N
-
 from mathics.core.expression import Expression
 from mathics.core.symbols import (
     Symbol,