+++ mathics/builtin/system.py
@@ -19,9 +19,7 @@ from mathics.core.atoms import (
     Real,
     String,
 )
-from mathics.core.symbols import (
-    SymbolList,
-)
+from mathics.core.symbols import SymbolList
 from mathics.core.systemsymbols import (
     SymbolFailed,
     SymbolRule,